src/requirements.txt
streamlit
pandas
numpy
pytz
matplotlib
plotly
scikit-learn
hopsworks
python-dotenv
joblib
catboost
requests